+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -249,7 +249,10 @@ public class SentryOptions {
   /** list of scope observers */
   private final @NotNull List<IScopeObserver> observers = new ArrayList<>();
 
-  /** Enable the Java to NDK Scope sync */
+  /**
+   * Enable the Java to NDK Scope sync. The default value for sentry-java is disabled and enabled
+   * for sentry-android.
+   */
   private boolean enableScopeSync;
 
   /**